+++ src/main/java/stein197/bf/Parser.java
@@ -1,5 +1,6 @@
 package stein197.bf;
 
+import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -14,14 +15,42 @@ class Parser {
 	 * @param data String to parse.
 	 * @return Sequence of brainfuck commands.
 	 */
-	public static Command[] parse(String data) {
+	public static SourceInfo parse(String data) throws ParseException {
 		char[] charArray = data.toCharArray();
 		List<Command> result = new ArrayList<Command>(data.length());
-		for (char character : charArray) {
+		int minCapacity = 0;
+		List<Integer> openCmdPositionList = new ArrayList<Integer>();
+		for (int i = 0; i < charArray.length; i++) {
+			var character = charArray[i];
 			var cmd = Command.from(character);
-			if (cmd != null)
-				result.add(cmd);
+			if (cmd == null)
+				continue;
+			switch (cmd) {
+				case NEXT: {
+					minCapacity++;
+					break;
+				}
+				case PREV: {
+					minCapacity--;
+					break;
+				}
+				case OPEN: {
+					openCmdPositionList.add(i);
+					break;
+				}
+				case CLOSE: {
+					if (openCmdPositionList.size() == 0)
+						throw new ParseException("Close command \"]\" has no corresponding open command", i);
+					openCmdPositionList.remove(openCmdPositionList.size() - 1);
+					break;
+				}
+			}
+			if (minCapacity < 0)
+				throw new ParseException("Source code defines negative pointer offset", i);
+			result.add(cmd);
 		}
-		return result.toArray(new Command[result.size()]);
+		if (openCmdPositionList.size() > 0)
+			throw new ParseException("Open command \"[\" has no corresponding close command", openCmdPositionList.get(openCmdPositionList.size() - 1));
+		return new SourceInfo(result.toArray(new Command[result.size()]), minCapacity + 1);
 	}
 }