+++ src/main/java/stein197/bf/SourceInfo.java
@@ -1,0 +1,3 @@
+package stein197.bf;
+
+record SourceInfo(Command[] commands, int minCapacity) {}