+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # CHANGELOG
 ## [Unreleased]
-- Check for redundant open/close brackets at parsing time
 - Populate source code with docblocks
 - Add unit tests/coverage
 - Add compiler from/to files and own compiled code format
 - Add infinite VM capacity
+
+## [0.2.0](../../compare/0.1.0...0.2.0) - 2021-08-15
+### Added
+- Check for redundant open/close commands at parsing time
+- Check for negative jump offset at parsing time
+- Capacity is automatically detected for VM at parsing time
 
 ## [0.1.0](../../tree/0.1.0) - 2021-08-15
 Release